refactor on_file_done logic
