+++ src/extraction/main.py
@@ -1,18 +1,48 @@
 """This module contains the main pipeline for the boreholes data extraction."""
 
 import logging
+from collections.abc import Callable
+from functools import partial
 from pathlib import Path
 
 import click
 from dotenv import load_dotenv
 
 from core.benchmark_utils import configure_logging
+from extraction.core.extract import ExtractionResult
 from extraction.evaluation.benchmark.spec import parse_benchmark_spec
 from extraction.runner import ExtractionBenchmarkRunner, ExtractionOptions, ExtractionPipelineRunner
+from extraction.utils.benchmark_utils import on_file_done
 from swissgeol_doc_processing.utils.file_utils import get_data_path
 
 load_dotenv()
 logger = logging.getLogger(__name__)
+
+
+def make_callback_factory(
+    *,
+    write_csv: bool,
+    skip_draw_predictions: bool,
+    draw_lines: bool,
+    draw_tables: bool,
+    draw_strip_logs: bool,
+) -> Callable[[Path, Path], Callable[[ExtractionResult], None]]:
+    """Return a factory that builds an on_file_done callback for a given output and input directory."""
+
+    def factory(out_directory: Path, input_directory: Path) -> Callable[[ExtractionResult], None]:
+        return partial(
+            on_file_done,
+            out_directory=out_directory,
+            input_directory=input_directory,
+            write_csv=write_csv,
+            input_is_file=input_directory.is_file(),
+            skip_draw_predictions=skip_draw_predictions,
+            draw_lines=draw_lines,
+            draw_tables=draw_tables,
+            draw_strip_logs=draw_strip_logs,
+        )
+
+    return factory
 
 
 def common_options(f):
@@ -98,7 +128,7 @@ def common_options(f):
         "--resume",
         is_flag=True,
         default=False,
-        help="Whether to resume extraction. Deaults to False.",
+        help="Whether to resume extraction. Defaults to False.",
     )(f)
     return f
 
@@ -135,23 +165,23 @@ def click_pipeline(
     # Setup logging (same for all)
     configure_logging()
 
+    factory = make_callback_factory(
+        write_csv=csv,
+        skip_draw_predictions=skip_draw_predictions,
+        draw_lines=draw_lines,
+        draw_tables=draw_tables,
+        draw_strip_logs=draw_strip_logs,
+    )
+
     # --- Multi-benchmark mode ---
     if benchmarks:
         specs = [parse_benchmark_spec(b) for b in benchmarks]
-
         ExtractionBenchmarkRunner(
             benchmarks=specs,
             multi_root=out_directory,
             resume=resume,
-            options=ExtractionOptions(
-                skip_draw_predictions=skip_draw_predictions,
-                draw_lines=draw_lines,
-                draw_tables=draw_tables,
-                draw_strip_logs=draw_strip_logs,
-                csv=csv,
-                matching_analytics=matching_analytics,
-                part=part,
-            ),
+            options=ExtractionOptions(matching_analytics=matching_analytics, part=part),
+            on_file_done_factory=factory,
         ).run()
     # --- Single-benchmark mode ---
     else:
@@ -165,15 +195,8 @@ def click_pipeline(
             ground_truth_path=ground_truth_path,
             out_directory=out_directory,
             metadata_path=metadata_path,
-            options=ExtractionOptions(
-                skip_draw_predictions=skip_draw_predictions,
-                draw_lines=draw_lines,
-                draw_tables=draw_tables,
-                draw_strip_logs=draw_strip_logs,
-                csv=csv,
-                matching_analytics=matching_analytics,
-                part=part,
-            ),
+            options=ExtractionOptions(matching_analytics=matching_analytics, part=part),
+            on_file_done=factory(out_directory, input_directory),
         ).execute()
 
 
@@ -194,6 +217,13 @@ def click_pipeline_metadata(
     matching_analytics: bool = False,
 ):
     """Run only the metadata part of the pipeline."""
+    factory = make_callback_factory(
+        write_csv=csv,
+        skip_draw_predictions=skip_draw_predictions,
+        draw_lines=draw_lines,
+        draw_tables=draw_tables,
+        draw_strip_logs=draw_strip_logs,
+    )
     ExtractionPipelineRunner(
         predictions_path=predictions_path,
         resume=bool(resume),
@@ -201,15 +231,8 @@ def click_pipeline_metadata(
         ground_truth_path=ground_truth_path,
         out_directory=out_directory,
         metadata_path=metadata_path,
-        options=ExtractionOptions(
-            skip_draw_predictions=skip_draw_predictions,
-            draw_lines=draw_lines,
-            draw_tables=draw_tables,
-            draw_strip_logs=draw_strip_logs,
-            csv=csv,
-            matching_analytics=matching_analytics,
-            part="metadata",
-        ),
+        options=ExtractionOptions(matching_analytics=matching_analytics, part="metadata"),
+        on_file_done=factory(out_directory, input_directory),
     ).execute()
